ESC exits second layer; mouse keys are regular hjkl when modifiers down
This also fixes an outlying deficiency in track_key -- now, if you track multiple
source keycodes to the same modifier pointer, it will correctly track that
several mods are held concurrently, and only state that the mod is up if all
source keycodes are released. This is implemented by using an int counter
instead of a bool flag.